+++ api/src/main/java/com/projectspring/api/Repositories/UserRepositories.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.projectspring.api.Models.UserEntities;
 
-public interface UserRepositories extends JpaRepository<UserEntities, Long> {
+public interface UserRepositories extends JpaRepository<UserEntities, Integer> {
     
     UserEntities findByUsername(String username);
 }
